Time: 163 ms (77.21%) | Memory: 19.3 MB (79.13%) - LeetSync

+++ 552-student-attendance-record-ii/student-attendance-record-ii.cpp
@@ -20,7 +20,7 @@ public:
         c = solve(i+1, abs_count, 0);
         c%=MOD;
 
-        return dp[i][abs_count][consec_late_count] = (((long long)a+b)%MOD + c)%MOD;
+        return dp[i][abs_count][consec_late_count] = ((a+b)%MOD + c)%MOD;
     }
 
     int checkRecord(int n) {
